Fixed camera issues. And orthographic camera actually works great??!
Re:last commit. The strategy was to build a whole movement BP from scratch. Made the cube collider the root, then attached a floating pawn movement component, and copied over the event graph. At first collisions did not cooperate and then played around with the values according to previous implementations and then it suddenly worked. Not sure what was changed or if it was a case of the editor not refreshing correctly. But it works, so for all intents and purposes the BP can be copied.